+++ routes-project/app/Repository/RouteRepository.php
@@ -3,6 +3,7 @@
 namespace App\Repository;
 
 use App\Models\Route;
+use App\Repository\Abstract\RepositoryAbstract;
 use App\Repository\Interface\IRepository;
 use Exception;
 
@@ -11,7 +12,7 @@ use Illuminate\Support\Facades\DB;
  * @author Nabil Leon Alvarez <@nalleon>
  * @author Pedro Martin Escuela <@PeterMartEsc>
  */
-class RouteRepository implements IRepository {
+class RouteRepository extends RepositoryAbstract implements IRepository {
 
     /**
      * Default constructor of the repository
@@ -19,45 +20,46 @@ class RouteRepository implements IRepository {
     public function __construct(){}
 
     /**
-     * Function to find all routes 
+     * Function to find all routes
      */
     public function findAll(): array{
         $list = [];
         try {
-            $routesMysql = Route::on("mysql")->get();
+            $routesMysql = Route::on($this->connectionMySql)->get();
             $list = $routesMysql->toArray();
-    
+
         } catch (\Exception $e) {
-            $routesSqlite = Route::on("sqlite")->get();
+            $routesSqlite = Route::on($this->connectionSqlite)->get();
             $list = $routesSqlite->toArray();
         }
-    
+
         return $list;
     }
-    
+
     /**
-     * Function to add an route 
+     * Function to add an route
      */
     public function save($p): object | null{
         $result = null;
         try {
-            $p->setConnection("mysql")->save();
+            $p->setConnection($this->connectionMySql)->save();
             $p->refresh();
             $result = $p;
 
-            $pSqlite = new Route();
-            $pSqlite->id = $p->id;
-            $pSqlite->title = $p->title;
-            $pSqlite->location = $p->location;
-            $pSqlite->distance = $p->distance;
-            $pSqlite->date_route = $p->date_route;
-            $pSqlite->difficulty = $p->difficulty;
-            $pSqlite->pets_allowed = $p->pets_allowed;
-            $pSqlite->vehicle_needed = $p->vehicle_needed;
-            $pSqlite->description = $p->description;
-            $pSqlite->user_id = $p->user_id;
-
-            $pSqlite->setConnection("sqlite")->save();
+            if(!app()->runningUnitTests()){
+                $pSqlite = new Route();
+                $pSqlite->id = $p->id;
+                $pSqlite->title = $p->title;
+                $pSqlite->location = $p->location;
+                $pSqlite->distance = $p->distance;
+                $pSqlite->date_route = $p->date_route;
+                $pSqlite->difficulty = $p->difficulty;
+                $pSqlite->pets_allowed = $p->pets_allowed;
+                $pSqlite->vehicle_needed = $p->vehicle_needed;
+                $pSqlite->description = $p->description;
+                $pSqlite->user_id = $p->user_id;
+                $pSqlite->setConnection($this->connectionSqlite)->save();
+            }
         } catch (\Exception $e) {
             throw new Exception($e->getMessage());
         }
@@ -76,10 +78,10 @@ class RouteRepository implements IRepository {
         //dd($lastQuery);
 
         try{
-            $pToFind = Route::on("mysql")->where("id", $id)->first();
+            $pToFind = Route::on($this->connectionMySql)->where("id", $id)->first();
         }catch(Exception $e){
             echo $e->getMessage();
-            $pToFind = Route::on("sqlite")->where("id", $id)->first();
+            $pToFind = Route::on($this->connectionSqlite)->where("id", $id)->first();
         }
 
         return $pToFind;
@@ -96,10 +98,10 @@ class RouteRepository implements IRepository {
         //dd($lastQuery);
 
         try{
-            $pToFind = Route::on("mysql")->where("title", $uniqueKey)->first();
+            $pToFind = Route::on($this->connectionMySql)->where("title", $uniqueKey)->first();
         }catch(Exception $e){
             echo $e->getMessage();
-            $pToFind = Route::on("sqlite")->where("title", $uniqueKey)->first();
+            $pToFind = Route::on($this->connectionSqlite)->where("title", $uniqueKey)->first();
         }
 
         return $pToFind;
@@ -110,10 +112,10 @@ class RouteRepository implements IRepository {
      */
     public function update($p): bool {
         $updated = false;
-    
+
         try {
-            $pUpdate = Route::on("mysql")->find($p->id);
-    
+            $pUpdate = Route::on($this->connectionMySql)->find($p->id)->first();
+
             if ($pUpdate) {
                 $pUpdate->id = $p->id;
                 $pUpdate->title = $p->title;
@@ -128,28 +130,30 @@ class RouteRepository implements IRepository {
                 $pUpdate->save();
                 $updated = true;
             }
-    
-            $pUpdateSqlite = Route::on("sqlite")->find($p->id);
-    
-            if ($pUpdateSqlite) {
-                $pUpdateSqlite->id = $p->id;
-                $pUpdateSqlite->title = $p->title;
-                $pUpdateSqlite->location = $p->location;
-                $pUpdateSqlite->distance = $p->distance;
-                $pUpdateSqlite->date_route = $p->date_route;
-                $pUpdateSqlite->difficulty = $p->difficulty;
-                $pUpdateSqlite->pets_allowed = $p->pets_allowed;
-                $pUpdateSqlite->vehicle_needed = $p->vehicle_needed;
-                $pUpdateSqlite->description = $p->description;
-                $pUpdateSqlite->user_id = $p->user_id;
-                $pUpdateSqlite->save();
-                $updated = true;
+
+
+            if(!app()->runningUnitTests()){
+                $pUpdateSqlite = Route::on($this->connectionSqlite)->find($p->id)->first();
+                if ($pUpdateSqlite) {
+                    $pUpdateSqlite->id = $p->id;
+                    $pUpdateSqlite->title = $p->title;
+                    $pUpdateSqlite->location = $p->location;
+                    $pUpdateSqlite->distance = $p->distance;
+                    $pUpdateSqlite->date_route = $p->date_route;
+                    $pUpdateSqlite->difficulty = $p->difficulty;
+                    $pUpdateSqlite->pets_allowed = $p->pets_allowed;
+                    $pUpdateSqlite->vehicle_needed = $p->vehicle_needed;
+                    $pUpdateSqlite->description = $p->description;
+                    $pUpdateSqlite->user_id = $p->user_id;
+                    $pUpdateSqlite->save();
+                    $updated = true;
+                }
             }
-    
+
         } catch (\Exception $e) {
             throw new Exception($e->getMessage());
         }
-    
+
         return $updated;
     }
 
@@ -159,22 +163,25 @@ class RouteRepository implements IRepository {
     public function delete($id): bool{
         $deleted = false;
         try {
-            $mySqlItem = Route::on("mysql")->find($id);
+            $mySqlItem = Route::on($this->connectionMySql)->find($id)->first();
             if ($mySqlItem) {
                 $mySqlItem->delete();
                 $deleted = true;
             }
-    
-            $sqliteItem = Route::on("sqlite")->find($id);
-            if ($sqliteItem) {
-                $sqliteItem->delete();
-                $deleted = true;
+
+
+            if(!app()->runningUnitTests()){
+                $sqliteItem = Route::on($this->connectionSqlite)->find($id)->first();
+                if ($sqliteItem) {
+                    $sqliteItem->delete();
+                    $deleted = true;
+                }
             }
-    
+
         } catch (\Exception $e) {
             throw new Exception($e->getMessage());
         }
-    
+
         return $deleted;
     }
 }
